requirements.txt
playwright>=1.40.0
pydantic>=2.0.0
aiohttp>=3.9.0
python-dotenv>=1.0.0
flask>=3.0.0
